requirement.txt
apt install \
python3-uvicorn \
python3-ldap3 \
python3-fastapi \ 
python3-dotenv \